More precise naming, relocated to .gui.importer
